+++ src/landing-page/src/utils/get-bot-config.ts
@@ -1,22 +1,7 @@
 import { GetItemCommand } from "@aws-sdk/client-dynamodb";
 import { marshall, unmarshall } from "@aws-sdk/util-dynamodb";
+import type { BotConfig, GetBotConfigResult } from "~/types";
 import { client } from ".";
-
-type BotConfig = {
-  guild_id: string;
-  action_id: string;
-  roles: Array<string>;
-};
-
-export type GetBotConfigResult =
-  | {
-      data: BotConfig;
-      error?: never;
-    }
-  | {
-      data?: null;
-      error: Error;
-    };
 
 // @REVIEW not sure where to store credentials
 const TABLE_NAME = process.env.REACT_APP_AWS_DYNAMODB_TABLE_NAME;
@@ -29,31 +14,33 @@ export const getBotConfig = async (guild_id: string) => {
     error: new Error("Data was never fetched"),
   };
 
+  let data;
+
   try {
-    const data = await client.send(
+    data = await client.send(
       new GetItemCommand({
         TableName: TABLE_NAME,
         Key: marshall(key),
       }),
     );
-
-    if (!data.Item) {
-      result = {
-        data: null,
-        error: new Error("There is no data for this guild in DB"),
-      };
-    }
-
-    if (data.Item) {
-      const item = unmarshall(data.Item);
-      result = { data: item as BotConfig, error: undefined };
-    }
   } catch (error) {
     result = {
       error: new Error("There is no data for this guild in DB", {
         cause: error as Error,
       }),
     };
+  }
+
+  if (!data?.Item) {
+    result = {
+      data: null,
+      error: new Error("There is no data for this guild in DB"),
+    };
+  }
+
+  if (data?.Item) {
+    const item = unmarshall(data?.Item);
+    result = { data: item as BotConfig, error: undefined };
   }
 
   return result;
