+++ src/types/bot-config.ts
@@ -1,0 +1,5 @@
+export type BotConfig = {
+  guild_id: string;
+  action_id: string;
+  roles: Array<string>;
+};
