+++ src/types/get-bot-config-result.ts
@@ -1,0 +1,11 @@
+import type { BotConfig } from ".";
+
+export type GetBotConfigResult =
+  | {
+      data: BotConfig;
+      error?: never;
+    }
+  | {
+      data?: null;
+      error: Error;
+    };
